Allows the developer to configure the frequency of the messages pinging the server to check for user notifications.
The main goal of this is to allow the developer to remove the clutter of mostly useless messages in order to see the rest of the information that Rails logs by defauls (e.g. the controller hit on each request, the parameters sent, et cetera.)

The default behavior is to ping the server every 30 seconds. By passing "?notification_seconds=120" the developer can request the ping to happen every 120 seconds instead.